requirements.txt
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
pushbullet.py==0.12.0
pycodestyle==2.10.0
python-dotenv==1.0.0
python-magic==0.4.27
requests==2.28.2
urllib3==1.26.15
websocket-client==1.5.1
